[CODEC-264]: Ensure hash128 maintains the sign extension bug.
The hash128(...) method was calling the public hash128x86(...) method
with an int seed when it should have called the private hash128x86(...)
with a long seed. Thus it did not have the sign extension error. The
internal method has been renamed to avoid a name clash with the public
API. The old hash128() behaviour is now restored as the seed is passed
to the hash method with implicit long conversion.